Implementing time independent function generators
